requirements.txt
Flask==2.0.1
flask-dapr==1.4.1
